HealthMate-main/backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
sqlalchemy
bcrypt
python-jose[cryptography]
email-validator
apscheduler
